<refactor> search 폴더 정리
